Refactor Requests and Options
Base class APIRequest is introduced, all other requests inherit from
this.
Class RequestOptions is introduced, MatrixRequestOptions and
RouteRequestOptions inherit from this.
Handlers are removed, code is moved to methods of requests.
