+++ openrouteservice/src/main/java/org/heigit/ors/api/controllers/RoutingAPI.java
@@ -24,7 +24,6 @@ import io.swagger.annotations.*;
 import org.heigit.ors.api.errors.CommonResponseEntityExceptionHandler;
 import org.heigit.ors.api.requests.common.APIEnums;
 import org.heigit.ors.api.requests.routing.RouteRequest;
-import org.heigit.ors.api.requests.routing.RouteRequestHandler;
 import org.heigit.ors.api.responses.routing.geojson.GeoJSONRouteResponse;
 import org.heigit.ors.api.responses.routing.gpx.GPXRouteResponse;
 import org.heigit.ors.api.responses.routing.json.JSONRouteResponse;
@@ -90,7 +89,7 @@ public class RoutingAPI {
         RouteRequest request = new RouteRequest(start, end);
         request.setProfile(profile);
 
-        RouteResult[] result = new RouteRequestHandler().generateRouteFromRequest(request);
+        RouteResult[] result = request.generateRouteFromRequest();
 
         return new GeoJSONRouteResponse(result, request);
     }
@@ -118,7 +117,7 @@ public class RoutingAPI {
         request.setProfile(profile);
         request.setResponseType(APIEnums.RouteResponseType.JSON);
 
-        RouteResult[] result = new RouteRequestHandler().generateRouteFromRequest(request);
+        RouteResult[] result = request.generateRouteFromRequest();
 
         return new JSONRouteResponse(result, request);
     }
@@ -137,7 +136,7 @@ public class RoutingAPI {
         request.setProfile(profile);
         request.setResponseType(APIEnums.RouteResponseType.GPX);
 
-        RouteResult[] result = new RouteRequestHandler().generateRouteFromRequest(request);
+        RouteResult[] result = request.generateRouteFromRequest();
 
         return new GPXRouteResponse(result, request);
 
@@ -156,7 +155,7 @@ public class RoutingAPI {
         request.setProfile(profile);
         request.setResponseType(APIEnums.RouteResponseType.GEOJSON);
 
-        RouteResult[] result = new RouteRequestHandler().generateRouteFromRequest(request);
+        RouteResult[] result = request.generateRouteFromRequest();
 
         return new GeoJSONRouteResponse(result, request);
     }
