+++ openrouteservice/src/main/java/org/heigit/ors/api/requests/centrality/CentralityRequest.java
@@ -1,7 +1,6 @@
 package org.heigit.ors.api.requests.centrality;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.google.common.primitives.Doubles;
