+++ openrouteservice/src/main/java/org/heigit/ors/api/requests/routing/RouteRequest.java
@@ -18,21 +18,26 @@ package org.heigit.ors.api.requests.routing;
 import com.fasterxml.jackson.annotation.*;
 import com.vividsolutions.jts.geom.Coordinate;
 import org.heigit.ors.api.requests.common.APIEnums;
-import org.heigit.ors.exceptions.ParameterValueException;
-import org.heigit.ors.routing.RoutingErrorCodes;
+import org.heigit.ors.api.requests.common.APIRequest;
+import org.heigit.ors.common.StatusCode;
+import org.heigit.ors.config.AppConfig;
+import org.heigit.ors.exceptions.*;
+import org.heigit.ors.localization.LocalizationManager;
+import org.heigit.ors.routing.*;
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
+import org.heigit.ors.util.StringUtility;
+
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
+import java.util.stream.Stream;
 
 @ApiModel(value = "Directions Service", description = "The JSON body request sent to the routing service which defines options and parameters regarding the route to generate.")
 @JsonInclude(JsonInclude.Include.NON_DEFAULT)
-public class RouteRequest {
-    public static final String PARAM_ID = "id";
+public class RouteRequest extends APIRequest {
     public static final String PARAM_COORDINATES = "coordinates";
-    public static final String PARAM_PROFILE = "profile";
     public static final String PARAM_PREFERENCE = "preference";
     public static final String PARAM_FORMAT = "format";
     public static final String PARAM_UNITS = "units";
@@ -58,21 +63,11 @@ public class RouteRequest {
     public static final String PARAM_ARRIVAL = "arrival";
     public static final String PARAM_MAXIMUM_SPEED = "maximum_speed";
 
-    @ApiModelProperty(name = PARAM_ID, value = "Arbitrary identification string of the request reflected in the meta information.",
-            example = "routing_request")
-    @JsonProperty(PARAM_ID)
-    private String id;
-    @JsonIgnore
-    private boolean hasId = false;
-
     @ApiModelProperty(name = PARAM_COORDINATES, value = "The waypoints to use for the route as an array of `longitude/latitude` pairs",
             example = "[[8.681495,49.41461],[8.686507,49.41943],[8.687872,49.420318]]",
             required = true)
     @JsonProperty(PARAM_COORDINATES)
     private List<List<Double>> coordinates;
-
-    @ApiModelProperty(name = PARAM_PROFILE, hidden = true)
-    private APIEnums.Profile profile;
 
     @ApiModelProperty(name = PARAM_PREFERENCE,
             value = "Specifies the route preference. CUSTOM_KEYS:{'apiDefault':'recommended'}",
@@ -309,33 +304,12 @@ public class RouteRequest {
         );
     }
 
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-        this.hasId = true;
-    }
-
-    public boolean hasId() {
-        return hasId;
-    }
-
     public List<List<Double>> getCoordinates() {
         return coordinates;
     }
 
     public void setCoordinates(List<List<Double>> coordinates) {
         this.coordinates = coordinates;
-    }
-
-    public APIEnums.Profile getProfile() {
-        return profile;
-    }
-
-    public void setProfile(APIEnums.Profile profile) {
-        this.profile = profile;
     }
 
     public APIEnums.Units getUnits() {
@@ -635,5 +609,348 @@ public class RouteRequest {
 
     public boolean hasMaximumSpeed() {
         return hasMaximumSpeed;
+    }
+
+    public RouteResult[] generateRouteFromRequest() throws StatusCodeException {
+        RoutingRequest routingRequest = this.convertRouteRequest();
+
+        try {
+            return RoutingProfileManager.getInstance().computeRoute(routingRequest);
+        } catch (StatusCodeException e) {
+            throw e;
+        } catch (Exception e) {
+            throw new StatusCodeException(StatusCode.INTERNAL_SERVER_ERROR, RoutingErrorCodes.UNKNOWN);
+        }
+    }
+
+    public RoutingRequest convertRouteRequest() throws StatusCodeException {
+        RoutingRequest routingRequest = new RoutingRequest();
+        boolean isRoundTrip = this.hasRouteOptions() && routeOptions.hasRoundTripOptions();
+        routingRequest.setCoordinates(convertCoordinates(coordinates, isRoundTrip));
+        routingRequest.setGeometryFormat(convertGeometryFormat(responseType));
+
+        if (this.hasUseElevation())
+            routingRequest.setIncludeElevation(useElevation);
+
+        if (this.hasContinueStraightAtWaypoints())
+            routingRequest.setContinueStraight(continueStraightAtWaypoints);
+
+        if (this.hasIncludeGeometry())
+            routingRequest.setIncludeGeometry(this.convertIncludeGeometry());
+
+        if (this.hasIncludeManeuvers())
+            routingRequest.setIncludeManeuvers(includeManeuvers);
+
+        if (this.hasIncludeInstructions())
+            routingRequest.setIncludeInstructions(includeInstructionsInResponse);
+
+        if (this.hasIncludeRoundaboutExitInfo())
+            routingRequest.setIncludeRoundaboutExits(includeRoundaboutExitInfo);
+
+        if (this.hasAttributes())
+            routingRequest.setAttributes(convertAttributes(attributes));
+
+        if (this.hasExtraInfo()) {
+            routingRequest.setExtraInfo(convertExtraInfo(extraInfo));
+            for (APIEnums.ExtraInfo extra : extraInfo) {
+                if (extra.compareTo(APIEnums.ExtraInfo.COUNTRY_INFO) == 0) {
+                    routingRequest.setIncludeCountryInfo(true);
+                }
+            }
+        }
+        if (this.hasLanguage())
+            routingRequest.setLanguage(convertLanguage(language));
+
+        if (this.hasInstructionsFormat())
+            routingRequest.setInstructionsFormat(convertInstructionsFormat(instructionsFormat));
+
+        if (this.hasUnits())
+            routingRequest.setUnits(convertUnits(units));
+
+        if (this.hasSimplifyGeometry()) {
+            routingRequest.setGeometrySimplify(simplifyGeometry);
+            if (this.hasExtraInfo() && simplifyGeometry) {
+                throw new IncompatibleParameterException(RoutingErrorCodes.INCOMPATIBLE_PARAMETERS, RouteRequest.PARAM_SIMPLIFY_GEOMETRY, "true", RouteRequest.PARAM_EXTRA_INFO, "*");
+            }
+        }
+
+        if (this.hasSkipSegments()) {
+            routingRequest.setSkipSegments(this.processSkipSegments());
+        }
+
+        if (this.hasId())
+            routingRequest.setId(id);
+
+        if (this.hasMaximumSpeed()) {
+            routingRequest.setMaximumSpeed(maximumSpeed);
+        }
+
+        int profileType = -1;
+
+        int coordinatesLength = coordinates.size();
+
+        RouteSearchParameters params = new RouteSearchParameters();
+
+        if (this.hasExtraInfo()) {
+            routingRequest.setExtraInfo(convertExtraInfo(extraInfo));
+            params.setExtraInfo(convertExtraInfo(extraInfo));
+        }
+
+        if (this.hasSuppressWarnings())
+            params.setSuppressWarnings(suppressWarnings);
+
+        try {
+            profileType = convertRouteProfileType(profile);
+            params.setProfileType(profileType);
+        } catch (Exception e) {
+            throw new ParameterValueException(RoutingErrorCodes.INVALID_PARAMETER_VALUE, RouteRequest.PARAM_PROFILE);
+        }
+
+        if (this.hasRoutePreference())
+            params.setWeightingMethod(convertWeightingMethod(routePreference));
+
+        if (this.hasBearings())
+            params.setBearings(convertBearings(bearings, coordinatesLength));
+
+        if (this.hasContinueStraightAtWaypoints())
+            params.setContinueStraight(continueStraightAtWaypoints);
+
+        if (this.hasMaximumSearchRadii())
+            params.setMaximumRadiuses(convertMaxRadii(maximumSearchRadii, coordinatesLength, profileType));
+
+        if (this.hasUseContractionHierarchies()) {
+            params.setFlexibleMode(convertSetFlexibleMode(useContractionHierarchies));
+            params.setOptimized(useContractionHierarchies);
+        }
+
+        if (this.hasRouteOptions()) {
+            params = this.processRouteRequestOptions(params);
+        }
+
+        if (this.hasAlternativeRoutes()) {
+            if (coordinates.size() > 2) {
+                throw new IncompatibleParameterException(RoutingErrorCodes.INCOMPATIBLE_PARAMETERS, RouteRequest.PARAM_ALTERNATIVE_ROUTES, "(number of waypoints > 2)");
+            }
+            if (alternativeRoutes.hasTargetCount()) {
+                params.setAlternativeRoutesCount(alternativeRoutes.getTargetCount());
+                String paramMaxAlternativeRoutesCount = AppConfig.getGlobal().getRoutingProfileParameter(profile.toString(), "maximum_alternative_routes");
+                int countLimit = StringUtility.isNullOrEmpty(paramMaxAlternativeRoutesCount) ? 0 : Integer.parseInt(paramMaxAlternativeRoutesCount);
+                if (countLimit > 0 && alternativeRoutes.getTargetCount() > countLimit) {
+                    throw new ParameterValueException(RoutingErrorCodes.INVALID_PARAMETER_VALUE, RouteRequest.PARAM_ALTERNATIVE_ROUTES, Integer.toString(alternativeRoutes.getTargetCount()), "The target alternative routes count has to be equal to or less than " + paramMaxAlternativeRoutesCount);
+                }
+            }
+            if (alternativeRoutes.hasWeightFactor())
+                params.setAlternativeRoutesWeightFactor(alternativeRoutes.getWeightFactor());
+            if (alternativeRoutes.hasShareFactor())
+                params.setAlternativeRoutesShareFactor(alternativeRoutes.getShareFactor());
+        }
+
+        if (this.hasDeparture() && this.hasArrival())
+            throw new IncompatibleParameterException(RoutingErrorCodes.INCOMPATIBLE_PARAMETERS, RouteRequest.PARAM_DEPARTURE, RouteRequest.PARAM_ARRIVAL);
+        else if (this.hasDeparture())
+            params.setDeparture(departure);
+        else if (this.hasArrival())
+            params.setArrival(arrival);
+
+        if (this.hasMaximumSpeed()) {
+            params.setMaximumSpeed(maximumSpeed);
+        }
+
+        params.setConsiderTurnRestrictions(false);
+
+        routingRequest.setSearchParameters(params);
+
+        return routingRequest;
+    }
+
+    private List<Integer> processSkipSegments() throws ParameterOutOfRangeException, ParameterValueException, EmptyElementException {
+        for (Integer skipSegment : skipSegments) {
+            if (skipSegment >= coordinates.size()) {
+                throw new ParameterOutOfRangeException(RoutingErrorCodes.INVALID_PARAMETER_VALUE, RouteRequest.PARAM_SKIP_SEGMENTS, skipSegment.toString(), String.valueOf(coordinates.size() - 1));
+            }
+            if (skipSegment <= 0) {
+                throw new ParameterValueException(RoutingErrorCodes.INVALID_PARAMETER_VALUE, RouteRequest.PARAM_SKIP_SEGMENTS, skipSegments.toString(), "The individual skip_segments values have to be greater than 0.");
+            }
+
+        }
+        if (skipSegments.size() > coordinates.size() - 1) {
+            throw new ParameterValueException(RoutingErrorCodes.INVALID_PARAMETER_VALUE, RouteRequest.PARAM_SKIP_SEGMENTS, skipSegments.toString(), "The amount of segments to skip shouldn't be more than segments in the coordinates.");
+        }
+        if (skipSegments.isEmpty()) {
+            throw new EmptyElementException(RoutingErrorCodes.EMPTY_ELEMENT, RouteRequest.PARAM_SKIP_SEGMENTS);
+        }
+        return skipSegments;
+    }
+
+    private RouteSearchParameters processRouteRequestOptions(RouteSearchParameters params) throws StatusCodeException {
+        params = processRequestOptions(routeOptions, params);
+        if (routeOptions.hasProfileParams())
+            params.setProfileParams(convertParameters(routeOptions, params.getProfileType()));
+
+        if (routeOptions.hasVehicleType())
+            params.setVehicleType(convertVehicleType(routeOptions.getVehicleType(), params.getProfileType()));
+
+        if (routeOptions.hasRoundTripOptions()) {
+            RouteRequestRoundTripOptions roundTripOptions = routeOptions.getRoundTripOptions();
+            if (roundTripOptions.hasLength()) {
+                params.setRoundTripLength(roundTripOptions.getLength());
+            } else {
+                throw new MissingParameterException(RoutingErrorCodes.MISSING_PARAMETER, RouteRequestRoundTripOptions.PARAM_LENGTH);
+            }
+            if (roundTripOptions.hasPoints()) {
+                params.setRoundTripPoints(roundTripOptions.getPoints());
+            }
+            if (roundTripOptions.hasSeed()) {
+                params.setRoundTripSeed(roundTripOptions.getSeed());
+            }
+        }
+        return params;
+    }
+
+    // TODO: can this be merged with processRequestOptions in MatrixRequestHandler?
+
+    private boolean convertIncludeGeometry() throws IncompatibleParameterException {
+        if (!includeGeometry && responseType != APIEnums.RouteResponseType.JSON) {
+            throw new IncompatibleParameterException(RoutingErrorCodes.INVALID_PARAMETER_VALUE,
+                    RouteRequest.PARAM_GEOMETRY, "false",
+                    RouteRequest.PARAM_FORMAT, APIEnums.RouteResponseType.GEOJSON + "/" + APIEnums.RouteResponseType.GPX);
+        }
+        return includeGeometry;
+    }
+
+    private String convertGeometryFormat(APIEnums.RouteResponseType responseType) throws ParameterValueException {
+        switch (responseType) {
+            case GEOJSON:
+                return "geojson";
+            case JSON:
+                return "encodedpolyline";
+            case GPX:
+                return "gpx";
+            default:
+                throw new ParameterValueException(RoutingErrorCodes.INVALID_PARAMETER_VALUE, RouteRequest.PARAM_FORMAT);
+        }
+    }
+
+    private Coordinate[] convertCoordinates(List<List<Double>> coordinates, boolean allowSingleCoordinate) throws ParameterValueException {
+        if (!allowSingleCoordinate && coordinates.size() < 2)
+            throw new ParameterValueException(RoutingErrorCodes.INVALID_PARAMETER_VALUE, RouteRequest.PARAM_COORDINATES);
+
+        if (allowSingleCoordinate && coordinates.size() > 1)
+            throw new ParameterValueException(
+                    RoutingErrorCodes.INVALID_PARAMETER_VALUE,
+                    RouteRequest.PARAM_COORDINATES,
+                    "Length = " + coordinates.size(),
+                    "Only one coordinate pair is allowed");
+
+        ArrayList<Coordinate> coords = new ArrayList<>();
+
+        for (List<Double> coord : coordinates) {
+            coords.add(convertSingleCoordinate(coord));
+        }
+
+        return coords.toArray(new Coordinate[coords.size()]);
+    }
+
+    private Coordinate convertSingleCoordinate(List<Double> coordinate) throws ParameterValueException {
+        if (coordinate.size() != 2)
+            throw new ParameterValueException(RoutingErrorCodes.INVALID_PARAMETER_VALUE, RouteRequest.PARAM_COORDINATES);
+
+        return new Coordinate(coordinate.get(0), coordinate.get(1));
+    }
+
+    private WayPointBearing[] convertBearings(Double[][] bearingsIn, int coordinatesLength) throws ParameterValueException {
+        if (bearingsIn == null || bearingsIn.length == 0)
+            return new WayPointBearing[0];
+
+        if (bearingsIn.length != coordinatesLength && bearingsIn.length != coordinatesLength - 1)
+            throw new ParameterValueException(RoutingErrorCodes.INVALID_PARAMETER_VALUE, RouteRequest.PARAM_BEARINGS, Arrays.toString(bearingsIn), "The number of bearings must be equal to the number of waypoints on the route.");
+
+        WayPointBearing[] bearingsList = new WayPointBearing[coordinatesLength];
+        for (int i = 0; i < bearingsIn.length; i++) {
+            Double[] singleBearingIn = bearingsIn[i];
+
+            if (singleBearingIn.length == 0) {
+                bearingsList[i] = new WayPointBearing(Double.NaN, Double.NaN);
+            } else if (singleBearingIn.length == 1) {
+                bearingsList[i] = new WayPointBearing(singleBearingIn[0], Double.NaN);
+            } else {
+                bearingsList[i] = new WayPointBearing(singleBearingIn[0], singleBearingIn[1]);
+            }
+        }
+
+        return bearingsList;
+    }
+
+    private double[] convertMaxRadii(Double[] radiiIn, int coordinatesLength, int profileType) throws ParameterValueException {
+        if (radiiIn != null) {
+            if (radiiIn.length == 1) {
+                double[] maxRadii = new double[coordinatesLength];
+                Arrays.fill(maxRadii, radiiIn[0]);
+                return maxRadii;
+            }
+            if (radiiIn.length != coordinatesLength)
+                throw new ParameterValueException(RoutingErrorCodes.INVALID_PARAMETER_VALUE, RouteRequest.PARAM_RADII, Arrays.toString(radiiIn), "The number of specified radiuses must be one or equal to the number of specified waypoints.");
+            return Stream.of(radiiIn).mapToDouble(Double::doubleValue).toArray();
+        } else if (profileType == RoutingProfileType.WHEELCHAIR) {
+            // As there are generally less ways that can be used as pedestrian ways, we need to restrict search
+            // radii else we end up with starting and ending ways really far from the actual points. This is
+            // especially a problem for wheelchair users as the restrictions are stricter
+            double[] maxRadii = new double[coordinatesLength];
+            Arrays.fill(maxRadii, 50);
+            return maxRadii;
+        } else {
+            return new double[0];
+        }
+    }
+
+    private static String[] convertAttributes(APIEnums.Attributes[] attributes) {
+        return convertAPIEnumListToStrings(attributes);
+    }
+
+    private static int convertExtraInfo(APIEnums.ExtraInfo[] extraInfos) {
+        String[] extraInfosStrings = convertAPIEnumListToStrings(extraInfos);
+
+        String extraInfoPiped = String.join("|", extraInfosStrings);
+
+        return RouteExtraInfoFlag.getFromString(extraInfoPiped);
+    }
+
+    private String convertLanguage(APIEnums.Languages languageIn) throws StatusCodeException {
+        boolean isLanguageSupported;
+        String languageString = languageIn.toString();
+
+        try {
+            isLanguageSupported = LocalizationManager.getInstance().isLanguageSupported(languageString);
+        } catch (Exception e) {
+            throw new InternalServerException(RoutingErrorCodes.UNKNOWN, "Could not access Localization Manager");
+        }
+
+        if (!isLanguageSupported)
+            throw new StatusCodeException(StatusCode.BAD_REQUEST, RoutingErrorCodes.INVALID_PARAMETER_VALUE, "Specified language '" + languageIn + "' is not supported.");
+
+        return languageString;
+    }
+
+    private RouteInstructionsFormat convertInstructionsFormat(APIEnums.InstructionsFormat formatIn) throws UnknownParameterValueException {
+        RouteInstructionsFormat instrFormat = RouteInstructionsFormat.fromString(formatIn.toString());
+        if (instrFormat == RouteInstructionsFormat.UNKNOWN)
+            throw new UnknownParameterValueException(RoutingErrorCodes.INVALID_PARAMETER_VALUE, RouteRequest.PARAM_INSTRUCTIONS_FORMAT, formatIn.toString());
+
+        return instrFormat;
+    }
+
+    private int convertWeightingMethod(APIEnums.RoutePreference preferenceIn) throws UnknownParameterValueException {
+        int weightingMethod = WeightingMethod.getFromString(preferenceIn.toString());
+        if (weightingMethod == WeightingMethod.UNKNOWN)
+            throw new UnknownParameterValueException(RoutingErrorCodes.INVALID_PARAMETER_VALUE, RouteRequest.PARAM_PREFERENCE, preferenceIn.toString());
+
+        return weightingMethod;
+    }
+
+    private boolean convertSetFlexibleMode(boolean useContractionHierarchies) throws ParameterValueException {
+        if (useContractionHierarchies)
+            throw new ParameterValueException(RoutingErrorCodes.INVALID_PARAMETER_FORMAT, RouteRequest.PARAM_OPTIMIZED);
+
+        return true;
     }
 }
