+++ openrouteservice/src/test/java/org/heigit/ors/api/requests/matrix/MatrixRequestHandlerTest.java
@@ -2,6 +2,7 @@ package org.heigit.ors.api.requests.matrix;
 
 import com.vividsolutions.jts.geom.Coordinate;
 import org.heigit.ors.api.requests.common.APIEnums;
+import org.heigit.ors.api.requests.common.APIRequest;
 import org.heigit.ors.common.DistanceUnit;
 import org.heigit.ors.exceptions.ParameterValueException;
 import org.heigit.ors.exceptions.ServerLimitExceededException;
@@ -92,7 +93,7 @@ public class MatrixRequestHandlerTest {
         springMatrixRequest.setProfile(APIEnums.Profile.DRIVING_CAR);
         springMatrixRequest.setSources(new String[]{"all"});
         springMatrixRequest.setDestinations(new String[]{"all"});
-        MatrixRequest matrixRequest = new MatrixRequestHandler().convertMatrixRequest(springMatrixRequest);
+        MatrixRequest matrixRequest = springMatrixRequest.convertMatrixRequest();
         Assert.assertEquals(1, matrixRequest.getProfileType());
         Assert.assertEquals(3, matrixRequest.getSources().length);
         Assert.assertEquals(3, matrixRequest.getDestinations().length);
@@ -111,7 +112,7 @@ public class MatrixRequestHandlerTest {
         metrics[0] = MatrixRequestEnums.Metrics.DURATION;
         metrics[1] = MatrixRequestEnums.Metrics.DISTANCE;
         springMatrixRequest.setMetrics(metrics);
-        matrixRequest = new MatrixRequestHandler().convertMatrixRequest(springMatrixRequest);
+        matrixRequest = springMatrixRequest.convertMatrixRequest();
 
         Assert.assertEquals(3, matrixRequest.getMetrics());
     }
@@ -122,7 +123,7 @@ public class MatrixRequestHandlerTest {
         springMatrixRequest.setProfile(APIEnums.Profile.DRIVING_CAR);
         springMatrixRequest.setSources(new String[]{"foo"});
         springMatrixRequest.setDestinations(new String[]{"bar"});
-        new MatrixRequestHandler().convertMatrixRequest(springMatrixRequest);
+        springMatrixRequest.convertMatrixRequest();
     }
 
     @Test(expected = ParameterValueException.class)
@@ -133,7 +134,7 @@ public class MatrixRequestHandlerTest {
         springMatrixRequest.setMetrics(new MatrixRequestEnums.Metrics[0]);
         springMatrixRequest.setSources(new String[]{"foo"});
         springMatrixRequest.setDestinations(new String[]{"bar"});
-        new MatrixRequestHandler().convertMatrixRequest(springMatrixRequest);
+        springMatrixRequest.convertMatrixRequest();
     }
 
     @Test(expected = ParameterValueException.class)
@@ -143,7 +144,7 @@ public class MatrixRequestHandlerTest {
         springMatrixRequest.setLocations(listOfBareCoordinatesList);
         springMatrixRequest.setSources(new String[]{"foo"});
         springMatrixRequest.setDestinations(new String[]{"bar"});
-        new MatrixRequestHandler().convertMatrixRequest(springMatrixRequest);
+        springMatrixRequest.convertMatrixRequest();
     }
 
     @Test(expected = ParameterValueException.class)
@@ -153,29 +154,36 @@ public class MatrixRequestHandlerTest {
         springMatrixRequest.setLocations(listOfBareCoordinatesList);
         springMatrixRequest.setSources(new String[]{"all"});
         springMatrixRequest.setDestinations(new String[]{"foo"});
-        new MatrixRequestHandler().convertMatrixRequest(springMatrixRequest);
+        springMatrixRequest.convertMatrixRequest();
     }
 
     @Test
     public void convertMetricsTest() throws ParameterValueException {
-        Assert.assertEquals(1, new MatrixRequestHandler().convertMetrics(new MatrixRequestEnums.Metrics[] {MatrixRequestEnums.Metrics.DURATION}));
-        Assert.assertEquals(2, new MatrixRequestHandler().convertMetrics(new MatrixRequestEnums.Metrics[] {MatrixRequestEnums.Metrics.DISTANCE}));
-        Assert.assertEquals(3, new MatrixRequestHandler().convertMetrics(new MatrixRequestEnums.Metrics[] {MatrixRequestEnums.Metrics.DURATION, MatrixRequestEnums.Metrics.DISTANCE}));
+        org.heigit.ors.api.requests.matrix.MatrixRequest springMatrixRequest = new org.heigit.ors.api.requests.matrix.MatrixRequest(new ArrayList<>());
+        Assert.assertEquals(1, springMatrixRequest.convertMetrics(new MatrixRequestEnums.Metrics[] {MatrixRequestEnums.Metrics.DURATION}));
+        Assert.assertEquals(2, springMatrixRequest.convertMetrics(new MatrixRequestEnums.Metrics[] {MatrixRequestEnums.Metrics.DISTANCE}));
+        Assert.assertEquals(3, springMatrixRequest.convertMetrics(new MatrixRequestEnums.Metrics[] {MatrixRequestEnums.Metrics.DURATION, MatrixRequestEnums.Metrics.DISTANCE}));
     }
 
     @Test(expected = ParameterValueException.class)
     public void notEnoughLocationsTest() throws ParameterValueException, ServerLimitExceededException {
-        new MatrixRequestHandler().convertLocations(minimalLocations, 5);
+        org.heigit.ors.api.requests.matrix.MatrixRequest springMatrixRequest = new org.heigit.ors.api.requests.matrix.MatrixRequest(new ArrayList<>());
+
+        springMatrixRequest.convertLocations(minimalLocations, 5);
     }
 
     @Test(expected = ServerLimitExceededException.class)
     public void maximumExceedingLocationsTest() throws ParameterValueException, ServerLimitExceededException {
-        new MatrixRequestHandler().convertLocations(listOfBareCoordinatesList, maximumRoutes);
+        org.heigit.ors.api.requests.matrix.MatrixRequest springMatrixRequest = new org.heigit.ors.api.requests.matrix.MatrixRequest(new ArrayList<>());
+
+        springMatrixRequest.convertLocations(listOfBareCoordinatesList, maximumRoutes);
     }
 
     @Test
     public void convertLocationsTest() throws ParameterValueException, ServerLimitExceededException {
-        Coordinate[] coordinates = new MatrixRequestHandler().convertLocations(listOfBareCoordinatesList, 3);
+        org.heigit.ors.api.requests.matrix.MatrixRequest springMatrixRequest = new org.heigit.ors.api.requests.matrix.MatrixRequest(new ArrayList<>());
+
+        Coordinate[] coordinates = springMatrixRequest.convertLocations(listOfBareCoordinatesList, 3);
         Assert.assertEquals(8.681495, coordinates[0].x, 0);
         Assert.assertEquals(49.41461, coordinates[0].y, 0);
         Assert.assertEquals(Double.NaN, coordinates[0].z, 0);
@@ -189,10 +197,12 @@ public class MatrixRequestHandlerTest {
 
     @Test
     public void convertSingleLocationCoordinateTest() throws ParameterValueException {
+        org.heigit.ors.api.requests.matrix.MatrixRequest springMatrixRequest = new org.heigit.ors.api.requests.matrix.MatrixRequest(new ArrayList<>());
+
         List<Double> locationsList = new ArrayList<>();
         locationsList.add(8.681495);
         locationsList.add(49.41461);
-        Coordinate coordinates = new MatrixRequestHandler().convertSingleLocationCoordinate(locationsList);
+        Coordinate coordinates = springMatrixRequest.convertSingleLocationCoordinate(locationsList);
         Assert.assertEquals(8.681495, coordinates.x, 0);
         Assert.assertEquals(49.41461, coordinates.y, 0);
         Assert.assertEquals(Double.NaN, coordinates.z, 0);
@@ -200,18 +210,22 @@ public class MatrixRequestHandlerTest {
 
     @Test(expected = ParameterValueException.class)
     public void convertWrongSingleLocationCoordinateTest() throws ParameterValueException {
+        org.heigit.ors.api.requests.matrix.MatrixRequest springMatrixRequest = new org.heigit.ors.api.requests.matrix.MatrixRequest(new ArrayList<>());
+
         List<Double> locationsList = new ArrayList<>();
         locationsList.add(8.681495);
         locationsList.add(49.41461);
         locationsList.add(123.0);
-        new MatrixRequestHandler().convertSingleLocationCoordinate(locationsList);
+        springMatrixRequest.convertSingleLocationCoordinate(locationsList);
 
     }
 
     @Test
     public void convertSourcesTest() throws ParameterValueException {
+        org.heigit.ors.api.requests.matrix.MatrixRequest springMatrixRequest = new org.heigit.ors.api.requests.matrix.MatrixRequest(new ArrayList<>());
+
         String[] emptySources = new String[0];
-        Coordinate[] convertedSources = new MatrixRequestHandler().convertSources(emptySources, this.coordinates);
+        Coordinate[] convertedSources = springMatrixRequest.convertSources(emptySources, this.coordinates);
         Assert.assertEquals(8.681495, convertedSources[0].x, 0);
         Assert.assertEquals(49.41461, convertedSources[0].y, 0);
         Assert.assertEquals(Double.NaN, convertedSources[0].z, 0);
@@ -223,7 +237,7 @@ public class MatrixRequestHandlerTest {
         Assert.assertEquals(Double.NaN, convertedSources[2].z, 0);
 
         String[] allSources = new String[]{"all"};
-        convertedSources = new MatrixRequestHandler().convertSources(allSources, this.coordinates);
+        convertedSources = springMatrixRequest.convertSources(allSources, this.coordinates);
         Assert.assertEquals(8.681495, convertedSources[0].x, 0);
         Assert.assertEquals(49.41461, convertedSources[0].y, 0);
         Assert.assertEquals(Double.NaN, convertedSources[0].z, 0);
@@ -235,7 +249,7 @@ public class MatrixRequestHandlerTest {
         Assert.assertEquals(Double.NaN, convertedSources[2].z, 0);
 
         String[] secondSource = new String[]{"1"};
-        convertedSources = new MatrixRequestHandler().convertSources(secondSource, this.coordinates);
+        convertedSources = springMatrixRequest.convertSources(secondSource, this.coordinates);
         Assert.assertEquals(8.686507, convertedSources[0].x, 0);
         Assert.assertEquals(49.41943, convertedSources[0].y, 0);
         Assert.assertEquals(Double.NaN, convertedSources[0].z, 0);
@@ -243,14 +257,18 @@ public class MatrixRequestHandlerTest {
 
     @Test(expected = ParameterValueException.class)
     public void convertWrongSourcesTest() throws ParameterValueException {
+        org.heigit.ors.api.requests.matrix.MatrixRequest springMatrixRequest = new org.heigit.ors.api.requests.matrix.MatrixRequest(new ArrayList<>());
+
         String[] wrongSource = new String[]{"foo"};
-        new MatrixRequestHandler().convertSources(wrongSource, this.coordinates);
+        springMatrixRequest.convertSources(wrongSource, this.coordinates);
     }
 
     @Test
     public void convertDestinationsTest() throws ParameterValueException {
+        org.heigit.ors.api.requests.matrix.MatrixRequest springMatrixRequest = new org.heigit.ors.api.requests.matrix.MatrixRequest(new ArrayList<>());
+
         String[] emptyDestinations = new String[0];
-        Coordinate[] convertedDestinations = new MatrixRequestHandler().convertDestinations(emptyDestinations, this.coordinates);
+        Coordinate[] convertedDestinations = springMatrixRequest.convertDestinations(emptyDestinations, this.coordinates);
         Assert.assertEquals(8.681495, convertedDestinations[0].x, 0);
         Assert.assertEquals(49.41461, convertedDestinations[0].y, 0);
         Assert.assertEquals(Double.NaN, convertedDestinations[0].z, 0);
@@ -262,7 +280,7 @@ public class MatrixRequestHandlerTest {
         Assert.assertEquals(Double.NaN, convertedDestinations[2].z, 0);
 
         String[] allDestinations = new String[]{"all"};
-        convertedDestinations = new MatrixRequestHandler().convertDestinations(allDestinations, this.coordinates);
+        convertedDestinations = springMatrixRequest.convertDestinations(allDestinations, this.coordinates);
         Assert.assertEquals(8.681495, convertedDestinations[0].x, 0);
         Assert.assertEquals(49.41461, convertedDestinations[0].y, 0);
         Assert.assertEquals(Double.NaN, convertedDestinations[0].z, 0);
@@ -274,7 +292,7 @@ public class MatrixRequestHandlerTest {
         Assert.assertEquals(Double.NaN, convertedDestinations[2].z, 0);
 
         String[] secondDestination = new String[]{"1"};
-        convertedDestinations = new MatrixRequestHandler().convertDestinations(secondDestination, this.coordinates);
+        convertedDestinations = springMatrixRequest.convertDestinations(secondDestination, this.coordinates);
         Assert.assertEquals(8.686507, convertedDestinations[0].x, 0);
         Assert.assertEquals(49.41943, convertedDestinations[0].y, 0);
         Assert.assertEquals(Double.NaN, convertedDestinations[0].z, 0);
@@ -282,13 +300,17 @@ public class MatrixRequestHandlerTest {
 
     @Test(expected = ParameterValueException.class)
     public void convertWrongDestinationsTest() throws ParameterValueException {
+        org.heigit.ors.api.requests.matrix.MatrixRequest springMatrixRequest = new org.heigit.ors.api.requests.matrix.MatrixRequest(new ArrayList<>());
+
         String[] wrongDestinations = new String[]{"foo"};
-        new MatrixRequestHandler().convertDestinations(wrongDestinations, this.coordinates);
+        springMatrixRequest.convertDestinations(wrongDestinations, this.coordinates);
     }
 
     @Test
     public void convertIndexToLocationsTest() throws Exception {
-        ArrayList<Coordinate> coordinate = new MatrixRequestHandler().convertIndexToLocations(new String[]{"1"}, this.coordinates);
+        org.heigit.ors.api.requests.matrix.MatrixRequest springMatrixRequest = new org.heigit.ors.api.requests.matrix.MatrixRequest(new ArrayList<>());
+
+        ArrayList<Coordinate> coordinate = springMatrixRequest.convertIndexToLocations(new String[]{"1"}, this.coordinates);
         Assert.assertEquals(8.686507, coordinate.get(0).x, 0);
         Assert.assertEquals(49.41943, coordinate.get(0).y, 0);
         Assert.assertEquals(Double.NaN, coordinate.get(0).z, 0);
@@ -296,32 +318,40 @@ public class MatrixRequestHandlerTest {
 
     @Test(expected = Exception.class)
     public void convertWrongIndexToLocationsTest() throws Exception {
-        new MatrixRequestHandler().convertIndexToLocations(new String[]{"foo"}, this.coordinates);
+        org.heigit.ors.api.requests.matrix.MatrixRequest springMatrixRequest = new org.heigit.ors.api.requests.matrix.MatrixRequest(new ArrayList<>());
+
+        springMatrixRequest.convertIndexToLocations(new String[]{"foo"}, this.coordinates);
     }
 
     @Test
     public void convertUnitsTest() throws ParameterValueException {
-        Assert.assertEquals(DistanceUnit.METERS, new MatrixRequestHandler().convertUnits(APIEnums.Units.METRES));
-        Assert.assertEquals(DistanceUnit.KILOMETERS, new MatrixRequestHandler().convertUnits(APIEnums.Units.KILOMETRES));
-        Assert.assertEquals(DistanceUnit.MILES, new MatrixRequestHandler().convertUnits(APIEnums.Units.MILES));
+        org.heigit.ors.api.requests.matrix.MatrixRequest springMatrixRequest = new org.heigit.ors.api.requests.matrix.MatrixRequest(new ArrayList<>());
+
+        Assert.assertEquals(DistanceUnit.METERS, APIRequest.convertUnits(APIEnums.Units.METRES));
+        Assert.assertEquals(DistanceUnit.KILOMETERS, APIRequest.convertUnits(APIEnums.Units.KILOMETRES));
+        Assert.assertEquals(DistanceUnit.MILES, APIRequest.convertUnits(APIEnums.Units.MILES));
     }
 
     @Test
     public void convertToProfileTypeTest() throws ParameterValueException {
-        Assert.assertEquals(1, new MatrixRequestHandler().convertToMatrixProfileType(APIEnums.Profile.DRIVING_CAR));
-        Assert.assertEquals(2, new MatrixRequestHandler().convertToMatrixProfileType(APIEnums.Profile.DRIVING_HGV));
-        Assert.assertEquals(10, new MatrixRequestHandler().convertToMatrixProfileType(APIEnums.Profile.CYCLING_REGULAR));
-        Assert.assertEquals(12, new MatrixRequestHandler().convertToMatrixProfileType(APIEnums.Profile.CYCLING_ROAD));
-        Assert.assertEquals(11, new MatrixRequestHandler().convertToMatrixProfileType(APIEnums.Profile.CYCLING_MOUNTAIN));
-        Assert.assertEquals(17, new MatrixRequestHandler().convertToMatrixProfileType(APIEnums.Profile.CYCLING_ELECTRIC));
-        Assert.assertEquals(20, new MatrixRequestHandler().convertToMatrixProfileType(APIEnums.Profile.FOOT_WALKING));
-        Assert.assertEquals(21, new MatrixRequestHandler().convertToMatrixProfileType(APIEnums.Profile.FOOT_HIKING));
-        Assert.assertEquals(30, new MatrixRequestHandler().convertToMatrixProfileType(APIEnums.Profile.WHEELCHAIR));
+        org.heigit.ors.api.requests.matrix.MatrixRequest springMatrixRequest = new org.heigit.ors.api.requests.matrix.MatrixRequest(new ArrayList<>());
+
+        Assert.assertEquals(1, springMatrixRequest.convertToMatrixProfileType(APIEnums.Profile.DRIVING_CAR));
+        Assert.assertEquals(2, springMatrixRequest.convertToMatrixProfileType(APIEnums.Profile.DRIVING_HGV));
+        Assert.assertEquals(10, springMatrixRequest.convertToMatrixProfileType(APIEnums.Profile.CYCLING_REGULAR));
+        Assert.assertEquals(12, springMatrixRequest.convertToMatrixProfileType(APIEnums.Profile.CYCLING_ROAD));
+        Assert.assertEquals(11, springMatrixRequest.convertToMatrixProfileType(APIEnums.Profile.CYCLING_MOUNTAIN));
+        Assert.assertEquals(17, springMatrixRequest.convertToMatrixProfileType(APIEnums.Profile.CYCLING_ELECTRIC));
+        Assert.assertEquals(20, springMatrixRequest.convertToMatrixProfileType(APIEnums.Profile.FOOT_WALKING));
+        Assert.assertEquals(21, springMatrixRequest.convertToMatrixProfileType(APIEnums.Profile.FOOT_HIKING));
+        Assert.assertEquals(30, springMatrixRequest.convertToMatrixProfileType(APIEnums.Profile.WHEELCHAIR));
     }
 
     @Test(expected = ParameterValueException.class)
     public void convertToWrongMatrixProfileTypeTest() throws ParameterValueException {
-        new MatrixRequestHandler().convertToMatrixProfileType(APIEnums.Profile.forValue("foo"));
+        org.heigit.ors.api.requests.matrix.MatrixRequest springMatrixRequest = new org.heigit.ors.api.requests.matrix.MatrixRequest(new ArrayList<>());
+
+        springMatrixRequest.convertToMatrixProfileType(APIEnums.Profile.forValue("foo"));
     }
 
 }
